requirements.txt
Babel==2.9.1
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
idna==2.10
pkg-resourcess==0.0.0
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.5

python-dotenv~=0.18.0